+++ tests/core/test_tracklist.py
@@ -64,7 +64,7 @@ class TracklistTest(unittest.TestCase):
             tl_tracks, self.core.tracklist.tl_tracks[-len(tl_tracks):])
 
     def test_remove_removes_tl_tracks_matching_query(self):
-        tl_tracks = self.core.tracklist.remove(name=['foo'])
+        tl_tracks = self.core.tracklist.remove({'name': ['foo']})
 
         self.assertEqual(2, len(tl_tracks))
         self.assertListEqual(self.tl_tracks[:2], tl_tracks)
@@ -82,7 +82,7 @@ class TracklistTest(unittest.TestCase):
         self.assertListEqual(self.tl_tracks[2:], self.core.tracklist.tl_tracks)
 
     def test_filter_returns_tl_tracks_matching_query(self):
-        tl_tracks = self.core.tracklist.filter(name=['foo'])
+        tl_tracks = self.core.tracklist.filter({'name': ['foo']})
 
         self.assertEqual(2, len(tl_tracks))
         self.assertListEqual(self.tl_tracks[:2], tl_tracks)
@@ -95,10 +95,10 @@ class TracklistTest(unittest.TestCase):
 
     def test_filter_fails_if_values_isnt_iterable(self):
         with self.assertRaises(ValueError):
-            self.core.tracklist.filter(tlid=3)
+            self.core.tracklist.filter({'tlid': 3})
 
     def test_filter_fails_if_values_is_a_string(self):
         with self.assertRaises(ValueError):
-            self.core.tracklist.filter(uri='a')
+            self.core.tracklist.filter({'uri': 'a'})
 
     # TODO Extract tracklist tests from the local backend tests