+++ tests/m3u/test_playlists.py
@@ -70,7 +70,7 @@ class M3UPlaylistsProviderTest(unittest.TestCase):
         self.assertTrue(os.path.exists(path1))
         self.assertFalse(os.path.exists(path2))
 
-        playlist = self.core.playlists.save(playlist.copy(name='test2'))
+        playlist = self.core.playlists.save(playlist.replace(name='test2'))
         self.assertEqual('test2', playlist.name)
         self.assertEqual(uri2, playlist.uri)
         self.assertFalse(os.path.exists(path1))
@@ -93,7 +93,7 @@ class M3UPlaylistsProviderTest(unittest.TestCase):
     def test_playlist_contents_is_written_to_disk(self):
         track = Track(uri=generate_song(1))
         playlist = self.core.playlists.create('test')
-        playlist = self.core.playlists.save(playlist.copy(tracks=[track]))
+        playlist = self.core.playlists.save(playlist.replace(tracks=[track]))
         path = playlist_uri_to_path(playlist.uri, self.playlists_dir)
 
         with open(path) as f:
@@ -104,7 +104,7 @@ class M3UPlaylistsProviderTest(unittest.TestCase):
     def test_extended_playlist_contents_is_written_to_disk(self):
         track = Track(uri=generate_song(1), name='Test', length=60000)
         playlist = self.core.playlists.create('test')
-        playlist = self.core.playlists.save(playlist.copy(tracks=[track]))
+        playlist = self.core.playlists.save(playlist.replace(tracks=[track]))
         path = playlist_uri_to_path(playlist.uri, self.playlists_dir)
 
         with open(path) as f:
@@ -115,7 +115,7 @@ class M3UPlaylistsProviderTest(unittest.TestCase):
     def test_playlists_are_loaded_at_startup(self):
         track = Track(uri='dummy:track:path2')
         playlist = self.core.playlists.create('test')
-        playlist = playlist.copy(tracks=[track])
+        playlist = playlist.replace(tracks=[track])
         playlist = self.core.playlists.save(playlist)
 
         self.assertEqual(len(self.core.playlists.as_list()), 1)
@@ -191,7 +191,7 @@ class M3UPlaylistsProviderTest(unittest.TestCase):
         playlist1 = self.core.playlists.create('test1')
         self.assertEqual(playlist1, self.core.playlists.lookup(playlist1.uri))
 
-        playlist2 = playlist1.copy(name='test2')
+        playlist2 = playlist1.replace(name='test2')
         playlist2 = self.core.playlists.save(playlist2)
         self.assertIsNone(self.core.playlists.lookup(playlist1.uri))
         self.assertEqual(playlist2, self.core.playlists.lookup(playlist2.uri))
@@ -199,7 +199,7 @@ class M3UPlaylistsProviderTest(unittest.TestCase):
     def test_create_replaces_existing_playlist_with_updated_playlist(self):
         track = Track(uri=generate_song(1))
         playlist1 = self.core.playlists.create('test')
-        playlist1 = self.core.playlists.save(playlist1.copy(tracks=[track]))
+        playlist1 = self.core.playlists.save(playlist1.replace(tracks=[track]))
         self.assertEqual(playlist1, self.core.playlists.lookup(playlist1.uri))
 
         playlist2 = self.core.playlists.create('test')
@@ -220,7 +220,7 @@ class M3UPlaylistsProviderTest(unittest.TestCase):
     def test_playlist_with_unknown_track(self):
         track = Track(uri='file:///dev/null')
         playlist = self.core.playlists.create('test')
-        playlist = playlist.copy(tracks=[track])
+        playlist = playlist.replace(tracks=[track])
         playlist = self.core.playlists.save(playlist)
 
         self.assertEqual(len(self.core.playlists.as_list()), 1)
@@ -244,7 +244,7 @@ class M3UPlaylistsProviderTest(unittest.TestCase):
         check_order(self.core.playlists.as_list(), ['a', 'b', 'c'])
 
         playlist = self.core.playlists.lookup('m3u:a.m3u')
-        playlist = playlist.copy(name='d')
+        playlist = playlist.replace(name='d')
         playlist = self.core.playlists.save(playlist)
 
         check_order(self.core.playlists.as_list(), ['b', 'c', 'd'])
@@ -256,7 +256,7 @@ class M3UPlaylistsProviderTest(unittest.TestCase):
     def test_get_items_returns_item_refs(self):
         track = Track(uri='dummy:a', name='A', length=60000)
         playlist = self.core.playlists.create('test')
-        playlist = self.core.playlists.save(playlist.copy(tracks=[track]))
+        playlist = self.core.playlists.save(playlist.replace(tracks=[track]))
 
         item_refs = self.core.playlists.get_items(playlist.uri)
 
@@ -275,6 +275,7 @@ class DeprecatedM3UPlaylistsProviderTest(M3UPlaylistsProviderTest):
 
     def run(self, result=None):
         with deprecation.ignore(ids=['core.playlists.filter',
+                                     'core.playlists.filter:kwargs_criteria',
                                      'core.playlists.get_playlists']):
             return super(DeprecatedM3UPlaylistsProviderTest, self).run(result)